+++ src/PortalBase/Web/Http/Contract/Psr7MessageCurlShellFormatterContract.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Heptacom\HeptaConnect\Portal\Base\Web\Http\Contract;
+
+/**
+ * Describes PSR-7 message formatter to generate cURL shell scripts.
+ */
+abstract class Psr7MessageCurlShellFormatterContract extends Psr7MessageFormatterContract
+{
+}